Basic changes made to move away from Assignment 3
