--- app/src/main/java/com/example/android/lifecycleweather/SettingsActivity.java
@@ -1,13 +1,0 @@
-package com.example.android.lifecycleweather;
-
-import android.os.Bundle;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-public class SettingsActivity extends AppCompatActivity {
-    @Override
-    protected void onCreate(Bundle savedInstanceState){
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_settings);
-    }
-}
