--- app/src/main/java/com/example/android/lifecycleweather/SettingsFragment.java
@@ -1,36 +1,0 @@
-package com.example.android.lifecycleweather;
-
-import android.content.SharedPreferences;
-import android.os.Bundle;
-
-import androidx.preference.EditTextPreference;
-import androidx.preference.PreferenceFragmentCompat;
-
-public class SettingsFragment extends PreferenceFragmentCompat implements SharedPreferences.OnSharedPreferenceChangeListener {
-    @Override
-    public void onCreatePreferences(Bundle savedInstanceState, String rootKey){
-        addPreferencesFromResource(R.xml.prefs);
-        EditTextPreference locationPref = findPreference(getString(R.string.pref_location_key));
-        locationPref.setSummary(locationPref.getText());
-    }
-
-    @Override
-    public void onSharedPreferenceChanged(SharedPreferences sharedPreferences, String key){
-        if (key.equals(getString(R.string.pref_location_key))){
-            EditTextPreference preference = findPreference(key);
-            preference.setSummary(preference.getText());
-        }
-    }
-
-    @Override
-    public void onResume(){
-        super.onResume();
-        getPreferenceScreen().getSharedPreferences().registerOnSharedPreferenceChangeListener(this);
-    }
-
-    @Override
-    public void onPause(){
-        super.onPause();
-        getPreferenceScreen().getSharedPreferences().unregisterOnSharedPreferenceChangeListener(this);
-    }
-}
