--- app/src/main/java/com/example/android/lifecycleweather/data/Status.java
@@ -1,7 +1,0 @@
-package com.example.android.lifecycleweather.data;
-
-public enum  Status {
-    LOADING,
-    ERROR,
-    SUCCESS
-}
