+++ app/src/main/java/com/example/android/tweetexplorer/ForecastItemDetailActivity.java
@@ -1,7 +1,7 @@
-package com.example.android.lifecycleweather;
+package com.example.android.tweetexplorer;
 
 import android.content.Intent;
-import androidx.core.app.ShareCompat;
+
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.preference.PreferenceManager;
 
@@ -13,9 +13,9 @@ import android.widget.ImageView;
 import android.widget.TextView;
 
 import com.bumptech.glide.Glide;
-import com.example.android.lifecycleweather.data.ForecastItem;
-import com.example.android.lifecycleweather.data.WeatherPreferences;
-import com.example.android.lifecycleweather.utils.OpenWeatherMapUtils;
+import com.example.android.tweetexplorer.data.ForecastItem;
+import com.example.android.tweetexplorer.data.WeatherPreferences;
+import com.example.android.tweetexplorer.utils.OpenWeatherMapUtils;
 
 import java.text.DateFormat;
