+++ app/src/main/java/com/example/android/tweetexplorer/SettingsActivity.java
@@ -1,0 +1,13 @@
+package com.example.android.tweetexplorer;
+
+import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+public class SettingsActivity extends AppCompatActivity {
+    @Override
+    protected void onCreate(Bundle savedInstanceState){
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_settings);
+    }
+}
