+++ app/src/main/java/com/example/android/tweetexplorer/data/Status.java
@@ -1,0 +1,7 @@
+package com.example.android.tweetexplorer.data;
+
+public enum  Status {
+    LOADING,
+    ERROR,
+    SUCCESS
+}
