+++ app/src/main/java/com/example/android/tweetexplorer/data/WeatherPreferences.java
@@ -1,4 +1,4 @@
-package com.example.android.lifecycleweather.data;
+package com.example.android.tweetexplorer.data;
 
 public class WeatherPreferences {
     private static final String DEFAULT_FORECAST_LOCATION = "Corvallis,OR,US";
