+++ app/src/main/res/values/colors.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="colorPrimary">#008577</color>
-    <color name="colorPrimaryDark">#00574B</color>
-    <color name="colorAccent">#D81B60</color>
+    <color name="colorPrimary">#0945e8</color>
+    <color name="colorPrimaryDark">#06267d</color>
+    <color name="colorAccent">#5076de</color>
 </resources>
