+++ app/src/test/java/com/example/android/tweetexplorer/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.android.lifecycleweather;
+package com.example.android.tweetexplorer;
 
 import org.junit.Test;
